fix(amplify-provider-awscloudformation): fixed deletion for large buckets (#3512)
* fix(amplify-provider-awscloudformation): fixed deletion for large bucket

Recursive call to collect and delete large number of objects

3447, 3451

* test(amplify-provider-awscloudformation): fix credential and warning

* test: moved to beforeeach

* test: trying to get this test running

* test: attempt2 to figure out the error

* test: test stuff

* fix(amplify-provider-awscloudformation): fixed deletion for large bucket

Recursive call to collect and delete large number of objects

3447, 3451

* test(amplify-provider-awscloudformation): fix credential and warning

* test: moved to beforeeach

* test: trying to get this test running

* test: attempt2 to figure out the error

* test: test stuff

* test: testing with aws-sdk directly

* test: added build to install

* test(amplify-provider-awscloudformation): test changes

* test: cleanup after changes to test my changes

* test(amplify-provider-awscloudformation): removed test

* fix(cli): skip delete if already exists

#3447, #3451

* refactor(amplify-provider-awscloudformation): deleteAllObjects refactor

+++ packages/amplify-provider-awscloudformation/src/aws-utils/aws-s3.js
@@ -26,10 +26,11 @@ class S3 {
       : projectDetails.teamProviderInfo[envName][providerName].DeploymentBucketName;
     s3Params.Bucket = projectBucket;
 
-    return this.s3
-      .putObject(s3Params)
-      .promise()
-      .then(() => projectBucket);
+    this.putFile(s3Params).then(() => projectBucket);
+  }
+
+  putFile(s3Params) {
+    return this.s3.putObject(s3Params).promise();
   }
 
   getFile(s3Params, envName = this.context.amplify.getEnvInfo().envName) {
@@ -52,7 +53,7 @@ class S3 {
     return this.ifBucketExists(bucketName).then(result => {
       if (!result) {
         this.context.print.warning(
-          'The specified S3 bucket to store the CloudFormation templates is not present. We are creating one for you....'
+          'The specified S3 bucket to store the CloudFormation templates is not present. We are creating one for you....',
         );
         this.context.print.warning(`Bucket name: ${bucketName}`);
 
@@ -67,47 +68,87 @@ class S3 {
       }
     });
   }
-
-  deleteAllObjects(bucketName) {
+  getAllObjectKeys(bucketName, continuationToken = null) {
     return new Promise((resolve, reject) => {
       this.s3
-        .listObjects({ Bucket: bucketName })
+        .listObjectsV2({ Bucket: bucketName, ContinuationToken: continuationToken })
         .promise()
         .then((result, lerr) => {
           if (lerr) {
             reject(lerr);
             return;
           }
-
-          const promises = result.Contents.map(r => this.s3.deleteObject({ Bucket: bucketName, Key: r.Key }).promise());
-          Promise.all(promises).then((results, errors) => {
-            if (!_.compact(errors).length) resolve();
-            else reject(errors);
+          const objects = result.Contents.map(r => {
+            return { Key: r.Key };
           });
+          if (result.IsTruncated) {
+            this.getAllObjectKeys(bucketName, result.NextContinuationToken).then((result, error) => {
+              if (error) {
+                reject(error);
+              } else {
+                resolve(objects.concat(result));
+              }
+            });
+          } else {
+            resolve(objects);
+          }
         });
+    });
+  }
+
+  deleteAllObjects(bucketName) {
+    return new Promise((resolve, reject) => {
+      this.getAllObjectKeys(bucketName).then((result, error) => {
+        if (error) reject(error);
+        const chunkedResult = _.chunk(result, 1000);
+
+        const deleteReq = chunkedResult
+          .map(res => {
+            return {
+              Bucket: bucketName,
+              Delete: {
+                Objects: res,
+                Quiet: false,
+              },
+            };
+          })
+          .map(delParams => this.s3.deleteObjects(delParams))
+          .map(delRequest => delRequest.promise());
+        Promise.all(deleteReq)
+          .then(resolve)
+          .catch(reject);
+      });
     });
   }
 
   deleteS3Bucket(bucketName) {
     return new Promise((resolve, reject) => {
-      this.deleteAllObjects(bucketName).then((result, err) => {
+      this.ifBucketExists(bucketName).then((exists, err) => {
         if (err) {
           reject(err);
-          return;
         }
-
-        this.s3
-          .deleteBucket({
-            Bucket: bucketName,
-          })
-          .promise()
-          .then((dresult, derr) => {
-            if (derr) {
-              reject(derr);
+        if (exists) {
+          this.deleteAllObjects(bucketName).then((result, err) => {
+            if (err) {
+              reject(err);
               return;
             }
-            resolve(dresult);
+            this.s3
+              .deleteBucket({
+                Bucket: bucketName,
+              })
+              .promise()
+              .then((dresult, derr) => {
+                if (derr) {
+                  reject(derr);
+                  return;
+                }
+                resolve(dresult);
+              });
           });
+        } else {
+          resolve();
+        }
       });
     });
   }
